refactor: move all fields from src/fields/core to src/fields
