+++ src/fields/FieldCheckbox.vue
