+++ src/fields/index.ts
@@ -1,23 +1,23 @@
 import { App, Component } from 'vue'
 import type { FieldPluginOptions } from '@/resources/types/generic'
 
-import FieldText from '@/fields/core/FieldText.vue'
-import FieldPassword from '@/fields/core/FieldPassword.vue'
-import FieldSelect from '@/fields/core/FieldSelect.vue'
-import FieldSelectNative from '@/fields/core/FieldSelectNative.vue'
-import FieldRadio from '@/fields/core/FieldRadio.vue'
-import FieldColor from '@/fields/core/FieldColor.vue'
-import FieldNumber from '@/fields/core/FieldNumber.vue'
-import FieldSwitch from '@/fields/core/FieldSwitch.vue'
-import FieldTextarea from '@/fields/core/FieldTextarea.vue'
-import FieldMask from '@/fields/core/FieldMask.vue'
-import FieldChecklist from '@/fields/core/FieldChecklist.vue'
-import FieldCheckbox from '@/fields/core/FieldCheckbox.vue'
-import FieldObject from '@/fields/core/FieldObject.vue'
+import FieldText from '@/fields/FieldText.vue'
+import FieldPassword from '@/fields/FieldPassword.vue'
+import FieldSelect from '@/fields/FieldSelect.vue'
+import FieldSelectNative from '@/fields/FieldSelectNative.vue'
+import FieldRadio from '@/fields/FieldRadio.vue'
+import FieldColor from '@/fields/FieldColor.vue'
+import FieldNumber from '@/fields/FieldNumber.vue'
+import FieldSwitch from '@/fields/FieldSwitch.vue'
+import FieldTextarea from '@/fields/FieldTextarea.vue'
+import FieldMask from '@/fields/FieldMask.vue'
+import FieldChecklist from '@/fields/FieldChecklist.vue'
+import FieldCheckbox from '@/fields/FieldCheckbox.vue'
+import FieldObject from '@/fields/FieldObject.vue'
 
-import FieldSubmit from '@/fields/core/FieldSubmit.vue'
-import FieldReset from '@/fields/core/FieldReset.vue'
-import FieldButton from '@/fields/core/FieldButton.vue'
+import FieldSubmit from '@/fields/FieldSubmit.vue'
+import FieldReset from '@/fields/FieldReset.vue'
+import FieldButton from '@/fields/FieldButton.vue'
 
 
 const fieldComponents = {
