+++ tests/unit/components/fields/FieldButton.spec.ts
@@ -2,9 +2,9 @@ import { mountFormGenerator, generatePropsSingleField } from '@test/_resources/u
 import { describe, it, expect, beforeAll } from 'vitest'
 import { mount, config } from '@vue/test-utils'
 
-import FieldButton from '@/fields/core/FieldButton.vue'
-import FieldPassword from '@/fields/core/FieldPassword.vue'
-import FieldCheckbox from '@/fields/core/FieldCheckbox.vue'
+import FieldButton from '@/fields/FieldButton.vue'
+import FieldPassword from '@/fields/FieldPassword.vue'
+import FieldCheckbox from '@/fields/FieldCheckbox.vue'
 
 beforeAll(() => {
   config.global.components = { FieldButton }
